WIP update to latest atom core APIs
Took an attempt at fixing the issues that break the todo-list package
from loading in atom. It turns out a good majority of the underlying
APIs that this package rely on have been deprecated. If you comment out
the 2 `@todo` sections, you can get the package to load in atom but
it’s still not functional. Wanted to check-in my work so someone else
might be able to take it the last mile.

I’m switching to using atom-imdone instead of this package because it’s
more feature rich anyway but I like the minimalism this package has
taken.